Fix: Move interface into Resource namespace
