+++ src/Serializer/Link/LinkCollection.php
@@ -5,9 +5,9 @@ namespace Refinery29\ApiOutput\Serializer\Link;
 use Exception;
 use Refinery29\ApiOutput\Resource\Link\Link;
 use Refinery29\ApiOutput\Resource\Link\LinkCollection as Input;
+use Refinery29\ApiOutput\Resource\TopLevelResource;
 use Refinery29\ApiOutput\Serializer\HasSerializer;
 use Refinery29\ApiOutput\Serializer\Serializer;
-use Refinery29\ApiOutput\TopLevelResource;
 
 class LinkCollection implements Serializer, TopLevelResource
 {
