+++ src/Serializer/Result.php
@@ -3,7 +3,7 @@
 namespace Refinery29\ApiOutput\Serializer;
 
 use Refinery29\ApiOutput\Resource\Result as Input;
-use Refinery29\ApiOutput\TopLevelResource;
+use Refinery29\ApiOutput\Resource\TopLevelResource;
 
 class Result implements Serializer, TopLevelResource
 {
